Separate BeltCipher to modes classes
