+++ src/main/java/by/bcrypto/bee2j/provider/cipher/BeltCipher.java
@@ -13,133 +13,6 @@ import java.util.Arrays;
 
 public class BeltCipher extends CipherSpi {
 
-    public static final class BeltECB extends BeltCipher {
-        public BeltECB() {
-            super("ECB");
-        }
-    }
-
-    public static final class BeltCBC extends BeltCipher {
-        public BeltCBC() {
-            super("CBC");
-        }
-        @Override
-        protected byte[] initState(byte[] key, byte[] iv) {
-            byte[] state = new byte[(int)bee2.beltCBC_keep()];
-            bee2.beltCBCStart(state, key, key.length, iv);
-            return state;
-        }
-        @Override
-        protected byte[] updateState(byte[] data, int op, byte[] state) {
-            byte[] result = data.clone();
-            if(op == 1)
-                bee2.beltCBCStepE(result, result.length, state);
-            else if (op == 2)
-                bee2.beltCBCStepD(result, result.length, state);
-            return result;
-        }
-    }
-
-    public static final class BeltCFB extends BeltCipher {
-        public BeltCFB() {
-            super("CFB");
-        }
-        @Override
-        protected byte[] initState(byte[] key, byte[] iv) {
-            byte[] state = new byte[(int)bee2.beltCFB_keep()];
-            bee2.beltCFBStart(state, key, key.length, iv);
-            return state;
-        }
-        @Override
-        protected byte[] updateState(byte[] data, int op, byte[] state) {
-            byte[] result = data.clone();
-            if(op == 1)
-                bee2.beltCFBStepE(result, result.length, state);
-            else if (op == 2)
-                bee2.beltCFBStepD(result, result.length, state);
-            return result;
-        }
-    }
-
-    public static final class BeltCTR extends BeltCipher {
-        public BeltCTR() {
-            super("CTR");
-        }
-        @Override
-        protected byte[] initState(byte[] key, byte[] iv) {
-            byte[] state = new byte[(int)bee2.beltCTR_keep()];
-            bee2.beltCTRStart(state, key, key.length, iv);
-            return state;
-        }
-        @Override
-        protected byte[] updateState(byte[] data, int op, byte[] state) {
-            byte[] result = data.clone();
-            if(op == 1)
-                bee2.beltCTRStepE(result, result.length, state);
-            else if (op == 2)
-                bee2.beltCTRStepE(result, result.length, state);
-            return result;
-        }
-    }
-
-    public static final class BeltMAC extends BeltCipher {
-        public BeltMAC() {
-            super("MAC");
-        }
-        @Override
-        protected byte[] initState(byte[] key, byte[] iv) {
-            byte[] state = new byte[(int)bee2.beltMAC_keep()];
-            bee2.beltMACStart(state, key, key.length);
-            return state;
-        }
-        @Override
-        protected byte[] updateState(byte[] data, int op, byte[] state) {
-            byte[] result = data.clone();
-            bee2.beltMACStepA(result, result.length, state);
-            return null;
-        }
-        @Override
-        protected byte[] finishState(byte[] state) {
-            byte[] mac = new byte[8];
-            bee2.beltMACStepG(mac, state);
-            return mac;
-        }
-    }
-
-    public static final class BeltDWP extends BeltCipher {
-        public BeltDWP() {
-            super("DWP");
-        }
-        @Override
-        protected byte[] initState(byte[] key, byte[] iv) {
-            byte[] state = new byte[(int)bee2.beltDWP_keep()];
-            bee2.beltDWPStart(state, key, key.length, iv);
-            return state;
-        }
-        @Override
-        protected byte[] updateState(byte[] data, int op, byte[] state) {
-            byte[] result = data.clone();
-            if(op == 1) {
-                bee2.beltDWPStepE(result, result.length, state);
-                bee2.beltDWPStepA(result, result.length, state);
-            } else if (op == 2) {
-                bee2.beltDWPStepA(result, result.length, state);
-                bee2.beltDWPStepD(result, result.length, state);
-            }   
-            return result;
-        }
-        @Override
-        protected byte[] finishState(byte[] state) {
-            byte[] mac = new byte[8];
-            bee2.beltDWPStepG(mac, state);
-            return mac;
-        }
-        @Override
-        protected void updateAAD(byte[] data, byte[] state) {
-            bee2.beltDWPStepI(data, data.length, state);
-        }
-    }
-
     enum BELT_MODE {
         ECB,
         CBC,
