+++ src/main/java/by/bcrypto/bee2j/provider/cipher/BeltECB.java
@@ -1,0 +1,7 @@
+package by.bcrypto.bee2j.provider.cipher;
+
+public final class BeltECB extends BeltCipher {
+    public BeltECB() {
+        super("ECB");
+    }
+}
